+++ es/index.js
@@ -1,6 +1,6 @@
 import { sortedUniq, permissionsFor, scopesFor, mergeObjects, randomColor, variableNameFor, areNotEmpty, idOf, flat, compact, uniq, pkg } from '@lykmapipo/common';
 import { rcFor, getString, getStringSet, isTest, getObject, apiVersion as apiVersion$1 } from '@lykmapipo/env';
-import { collectionNameOf, createSubSchema, copyInstance, createVarySubSchema, ObjectId, model, createSchema, Mixed, connect } from '@lykmapipo/mongoose-common';
+import { collectionNameOf, createSubSchema, copyInstance, createVarySubSchema, ObjectId, model, createSchema, Mixed, toObjectIds, areSameInstance, connect } from '@lykmapipo/mongoose-common';
 import { mount } from '@lykmapipo/express-common';
 import { Router, getFor, schemaFor, downloadFor, postFor, getByIdFor, patchFor, putFor, deleteFor, start as start$1 } from '@lykmapipo/express-rest-actions';
 import { map, zipObject, without, keys, mapValues, pick, includes, omit, omitBy, isEmpty, toUpper, find, forEach, merge, get, isFunction, isArray, flatMap, isMap, trim, uniqWith } from 'lodash';
@@ -2009,88 +2009,136 @@ PredefineSchema.statics.deleteByExtension = (optns, done) => {
 };
 
 /**
- * @name findRecursive
- * @function findRecursive
- * @description Find existing predefine recursively using given criteria
- * @param {object} criteria valid ancestor query options
+ * @name findChildren
+ * @function findChildren
+ * @description Find predefine children recursively using given criteria
+ * @param {object} criteria valid parent query options
  * @param {Function} done callback to invoke on success or error
  * @returns {object|Error} found predefines or error
  *
  * @author lally elias <lallyelias87@gmail.com>
  * @since 1.9.0
- * @version 0.1.0
+ * @version 0.2.0
  * @static
  * @example
  *
  * const criteria = { _id: ... };
- * Predefine.findRecursive(criteria, (error, results) => { ... });
+ * Predefine.findChildren(criteria, (error, results) => { ... });
  * // => [ Predefine{ ... }, ... ]
  *
  */
-PredefineSchema.statics.findRecursive = (criteria, done) => {
+PredefineSchema.statics.findChildren = (criteria, done) => {
   // TODO: use $graphLookUp
-  // TODO: improve n-queries
 
   // ref
   const Predefine = model(MODEL_NAME);
   let results = [];
 
   // collect results
-  const collectResults = (updates) => {
-    let collected = uniq([].concat(results).concat(updates));
-    collected = uniqWith(collected, (a, b) => {
-      return a.equals(b);
-    });
+  const collectResults = (...updates) => {
+    let collected = uniq([...results, ...updates]);
+    collected = uniqWith(collected, areSameInstance);
     return collected;
   };
 
-  // find ancestors
-  const findAncestors = (next) => {
-    return Predefine.find(criteria)
-      .setOptions({ autopopulate: false })
-      .exec((error, ancestors) => {
-        if (error) {
-          return next(error);
-        }
-        results = collectResults(ancestors);
-        const parentIds = uniq(map(ancestors, '_id'));
-        return next(null, parentIds);
-      });
-  };
-
   // find children by their parents
-  const findChildren = (parents, next) => {
-    const conditions = { 'relations.parent': { $in: [].concat(parents) } };
-    return Predefine.find(conditions)
-      .setOptions({ autopopulate: false })
-      .exec(next);
-  };
+  const findKids = (conditions, next) => {
+    // prepare query
+    const query = Predefine.find(conditions).setOptions({
+      autopopulate: false,
+    });
 
-  // find children recursively
-  const findChildrenRecursive = (parents, next) => {
-    // ensure parents
-    if (!isEmpty(parents)) {
-      return findChildren(parents, (error, children) => {
-        // back-off on error
-        if (error) {
-          return next(error);
+    // execute query
+    return query.exec((error, children) => {
+      // back-off on error
+      if (error) {
+        return next(error);
+      }
+
+      // continue find children
+      if (!isEmpty(children)) {
+        results = collectResults(...children);
+        const parentIds = uniq(toObjectIds(...children));
+        if (isEmpty(parentIds)) {
+          return next(null, results);
         }
-        // collect result
-        results = collectResults(children);
-        // continue find children
-        const parentIds = uniq(map(children, '_id'));
-        return findChildrenRecursive(parentIds, next);
-      });
-    }
-    // continue
-    return next(null, results);
-  };
+        return findKids({ 'relations.parent': { $in: parentIds } }, next);
+      }
 
-  // prepare tasks
-  const tasks = [findAncestors, findChildrenRecursive];
+      // continue
+      return next(null, results);
+    });
+  };
 
   // do find recursively
-  return waterfall(tasks, done);
+  return findKids(criteria, done);
+};
+
+/**
+ * @name findParents
+ * @function findParents
+ * @description Find predefine parent recursively using given criteria
+ * @param {object} criteria valid child query options
+ * @param {Function} done callback to invoke on success or error
+ * @returns {object|Error} found predefines or error
+ *
+ * @author lally elias <lallyelias87@gmail.com>
+ * @since 1.10.0
+ * @version 0.1.0
+ * @static
+ * @example
+ *
+ * const criteria = { _id: ... };
+ * Predefine.findParents(criteria, (error, results) => { ... });
+ * // => [ Predefine{ ... }, ... ]
+ *
+ */
+PredefineSchema.statics.findParents = (criteria, done) => {
+  // TODO: use $graphLookUp
+
+  // ref
+  const Predefine = model(MODEL_NAME);
+  let results = [];
+
+  // collect results
+  const collectResults = (...updates) => {
+    let collected = uniq([...results, ...updates]);
+    collected = uniqWith(collected, areSameInstance);
+    return collected;
+  };
+
+  // find parent by her children
+  const findAncestors = (conditions, next) => {
+    // prepare query
+    const query = Predefine.find(conditions).setOptions({
+      autopopulate: false,
+    });
+
+    // execute query
+    return query.exec((error, ancestors) => {
+      // back-off on error
+      if (error) {
+        return next(error);
+      }
+
+      // continue find parent
+      if (!isEmpty(ancestors)) {
+        results = collectResults(...ancestors);
+        const ancestorIds = uniq(map(ancestors, 'relations.parent'));
+        const parentIds = uniq(toObjectIds(...ancestorIds));
+        if (isEmpty(parentIds)) {
+          return next(null, results);
+        }
+        return findAncestors({ _id: { $in: parentIds } }, next);
+      }
+
+      // continue
+      return next(null, results);
+    });
+  };
+
+  // do find recursively
+  return findAncestors(criteria, done);
 };
 
 /* export predefine model */